+++ .travis/run_travis_job.sh
@@ -14,7 +14,7 @@ set -x
 if [ "$CHECK_RUST" = true ] 
 then
     # Install cargo-audit if it's not already.
-    cargo-audit -V || cargo install cargo-audit --force
+    cargo audit --version || cargo install cargo-audit --force
     # Install nightly rust version and clippy.
     rustup toolchain install $RUST_NIGHTLY_VERSION
     cargo +$RUST_NIGHTLY_VERSION clippy -V | grep $RUST_CLIPPY_VERSION || cargo +$RUST_NIGHTLY_VERSION install clippy --vers $RUST_CLIPPY_VERSION --force
@@ -47,5 +47,11 @@ then
     echo 'Rust checks are completed.'
 else
     cd "${TRAVIS_BUILD_DIR}"
-    ./run_all_tests.sh
+
+    ./run_all_tests.sh;
+    # Linux builds currently skip some tests, so only OSX builds should update code coverage report.
+    if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
+      # Upload the coverage report to coveralls
+      mvn org.eluder.coveralls:coveralls-maven-plugin:report
+    fi
 fi