+++ exonum-java-binding-cryptocurrency-demo/src/main/java/com/exonum/binding/cryptocurrency/transactions/TransferTx.java
@@ -22,6 +22,8 @@ import static com.exonum.binding.cryptocurrency.transactions.TransactionPrecondi
 import com.exonum.binding.common.crypto.PublicKey;
 import com.exonum.binding.common.message.BinaryMessage;
 import com.exonum.binding.cryptocurrency.CryptocurrencySchema;
+import com.exonum.binding.cryptocurrency.HistoryEntity;
+import com.exonum.binding.cryptocurrency.HistoryEntity.Builder;
 import com.exonum.binding.cryptocurrency.Wallet;
 import com.exonum.binding.storage.database.Fork;
 import com.exonum.binding.storage.indices.ProofMapIndexProxy;
@@ -32,7 +34,9 @@ import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
 import java.util.Objects;
 
-/** A transaction that transfers cryptocurrency between two wallets. */
+/**
+ * A transaction that transfers cryptocurrency between two wallets.
+ */
 public final class TransferTx extends AbstractTransaction implements Transaction {
 
   static final short ID = 2;
@@ -43,8 +47,7 @@ public final class TransferTx extends AbstractTransaction implements Transaction
   private final long sum;
 
   @VisibleForTesting
-  TransferTx(BinaryMessage message, long seed, PublicKey fromWallet, PublicKey toWallet,
-             long sum) {
+  TransferTx(BinaryMessage message, long seed, PublicKey fromWallet, PublicKey toWallet, long sum) {
     super(message);
     this.seed = seed;
     this.fromWallet = fromWallet;
@@ -59,8 +62,8 @@ public final class TransferTx extends AbstractTransaction implements Transaction
     checkTransaction(message, ID);
 
     try {
-      TxMessagesProtos.TransferTx messageBody =
-          TxMessagesProtos.TransferTx.parseFrom(message.getBody());
+      TxMessageProtos.TransferTx messageBody =
+          TxMessageProtos.TransferTx.parseFrom(message.getBody());
 
       long seed = messageBody.getSeed();
       PublicKey fromWallet = toPublicKey(messageBody.getFromWallet());
@@ -69,7 +72,7 @@ public final class TransferTx extends AbstractTransaction implements Transaction
 
       return new TransferTx(message, seed, fromWallet, toWallet, sum);
     } catch (InvalidProtocolBufferException e) {
-      throw new IllegalArgumentException("Invalid TxMessagesProtos.TransferTx buffer", e);
+      throw new IllegalArgumentException("Invalid TxMessageProtos.TransferTx buffer", e);
     }
   }
 
@@ -94,6 +97,16 @@ public final class TransferTx extends AbstractTransaction implements Transaction
       }
       wallets.put(fromWallet, new Wallet(from.getBalance() - sum));
       wallets.put(toWallet, new Wallet(to.getBalance() + sum));
+
+      HistoryEntity historyEntity = Builder.newBuilder()
+          .setSeed(seed)
+          .setWalletFrom(fromWallet)
+          .setWalletTo(toWallet)
+          .setAmount(sum)
+          .setTransactionHash(hash())
+          .build();
+      schema.walletHistory(fromWallet).add(historyEntity);
+      schema.walletHistory(toWallet).add(historyEntity);
     }
   }